Change fixmes for examples of tensors
Summary:
It seems that the fixme codes and messages changed and the examples weren't updated.

It confused me when I tried to run the example, as I thought that variadics needed perhaps some separate enabling of extensions (it was all fine, the examples for incorrect things were just using the wrong fixme codes).

Reviewed By: pradeep90

Differential Revision: D25940979

fbshipit-source-id: b8bd582c657e2c1448e9db92b4903b41f4d931db