Include callees of nested functions in the dependency graph.
Summary:
Problem: The previous diff added a nested function as the dependent of the outer function. However, Pysa still doesn't know about the dependencies of the nested function. This means that we may not propagate taint from the callee to the nested function.

This diff adds the above information to the dependency graph.

Note that forward and backward analysis on a nested function used a call graph created from that nested function. So, they worked as expected before this diff. However, the dependency graph created the call graph from the module source. And since we didn't call `Preprocessing.defines ~include_nested:true`, it meant that the dependency graph just considered the callees of non-nested functions.

Reviewed By: arthaud

Differential Revision: D25973149

fbshipit-source-id: 7d9abf9e461f4e4ee0f2145f147a4cf1b20dd4d5