Error legibly on non-literal expression used in Literal type
Summary:
For BET, I was looking into why we have unintelligible errors for things like:

```
from typing_extensions import Literal
x = 1
test: Literal[x] = {}
```
Undefined or invalid type [11]: Annotation `` is not defined as a type.

```
from typing_extensions import Literal

y = int

def foo() -> None:
    x = 1
    a: Literal[x] = 1
    b: Literal[y] = 1
```
Undefined or invalid type [11]: Annotation `foo` is not defined as a type.
Undefined or invalid type [11]: Annotation `` is not defined as a type.

----

See underlying diff summary for full root cause explanation.

This diff tackles introducing a readable error for literals that contain arbitrary names in them.

We are forced to do this check alongside pre-existing verification for untracked names (ie., Expression.Name objects), to avoid termination of analysis. Iif we find untracked names, we replace the annotation with Top to avoid trying to parse untracked names later on. We must assume literals with names inside are literal enums and parse the base of those name expressions during type checking.

When we throw the error, I think it covers more use cases / is more readable to the user to throw an Invalid type complaining that the expression is not a literal as expected (see above examples). It applies and is sensible even if the user was typo-ing the name of an enum.

Reviewed By: pradeep90

Differential Revision: D26026771

fbshipit-source-id: 403d3d31104a631c04b6ea10afcde56ccf96301f