properly propagate forward taint for super().__init__() in constructors
Summary:
Context:

We special case super().__init__() in the backwards analysis to ensure that the side effects to `self` get tracked. However, we previously overlooked the forward analysis portion of this. This means that sources defined in super() would be ignored. This diff centralizes how we determine super() and augments the forward analysis to essentially do what the backwards analysis does.

Reviewed By: arthaud

Differential Revision: D26115967

fbshipit-source-id: ea2daee67559b23e7d325c111abbb607d5826c86