Do not crash entire statistics job on libcst module fail
Summary:
5 projects in fbcode are currently failing to log any coverage statistics because libcst is recursing forever in them.

For now, let's log a warning on the modules that libcst can't parse, and finish the coverage pass for the rest of the project so numbers get logged to scuba and the coverage UI isn't useless for them.

Note: follow ups (1) debug libcst infinite recursion error, (2) if we can't fix this in all cases, we can also log modules that we failed to gather info on and make those statistics job return finished with warnings to make the situation more visible.

{F367494253}

Reviewed By: dkgi

Differential Revision: D26166290

fbshipit-source-id: 1b6e003bc99817ae0fa4e033c8c8c707e46ad88d