add a python script to explore pysa models and issues
Summary:
This patch implements a simple python script that provides utility commands to explore a taint output file. This has many advantages compared to sapp explorer: this will show TITO information for models, as well as sources and sinks that are not part of a trace. The sapp explorer reads from a sapp database, which removes sources and sinks that are not present in traces, to save space.

To use the script, simply run `python3 -i scripts/explore_pysa_models.py`, and see the help.

Reviewed By: pradeep90

Differential Revision: D26192531

fbshipit-source-id: ab19269261fb8d839562230a3c7743a8015ad69c