Refactor existing infer tests for readability
Summary:
Keep the obscure field indexing testing format for the few cases where it's actually verifying the structure/contents of the infer information output, but migrate the other tests that are just looking for error content functionality to something less horrible.

In the meantime, remove duplicate tests / ones that don't make sense.

killme

Reviewed By: dkgi

Differential Revision: D26294649

fbshipit-source-id: 5f16954e3fdaa99b8e9ed54d5719350cb25eec3f