classify annotation errors
Summary: This change makes it so that annotation errors always have the erroring annotation in context, and removes one major cause for unclassified error emissions. Went for a relatively flexible design of having the "why" be a string, since we have a wide variety of parsing errors for annotations in particular.

Reviewed By: arthaud

Differential Revision: D26285423

fbshipit-source-id: 4fb9a68c79f26e80bcfdc72f8747c36fd628852d