make ClassHierarchy.Untracked take a string instead a Type.t
Summary:
The basic premise here is that our ClassHierarchy.Untracked issue gets surfaced to users a decent amount, and has come up on GitHub during the type checking multiple times: https://github.com/facebook/pyre-check/issues/378 as the most recent example. OCaml isn't able to serialize Type.t's to user, but can do a better job with strings, especially when we're in the case of a `Scheduler.map_reduce`, where the backrtrace can be very hard to read.

For reference, here's what the exceptions have historically looked like:

```
AnalysisClassHierarchy.Untracked(_)
```

This diff changes the type of `ClassHierarchy.Untracked` to take a string, so that users at least know which type caused the issues. This isn't perfect, but can give us enough information to help do some basic debugging.

Consequences:

The good:

1. We get more readable error messages for some of the harder-to-debug errors we tend to find.

2. We've eliminated a lot of the Type.show's, since all we did with this exception type was display it.

The bad:

1. Analysis failure errors of different types will not be joined any more. I don't think this is a real concern.

2. The types of analysis failures are no longer dequalified. I think this is the real concern, but in this case, feel that the benefit of users being able to debug this outweighs the longer error message.

Reviewed By: grievejia

Differential Revision: D26323785

fbshipit-source-id: b63b626005c2b14246534373825b68f9a2e96784