Allow running just the taint integration tests.
Summary:
Problem: I want to run just the taint tests. If I change one source file and run `make test`, then it runs all Pyre and Pysa tests, which takes minutes.

We can't use the usual Dune trick for running a single test because the taint tests use special logic to create test cases from the .py files and compare the output to the .models files. `TestHelper.run_with_taint_models` runs each test in an ad hoc manner, which means we don't get a test id that we can run alone.

Solution: I'm wrapping the naked call to `TestHelper.run_with_taint_models` so that it is treated like any other unit test.

I tried unpacking the `TestHelper` test code to return a set of tests, but the above hack worked, so I'm leaving it at that. Nobody else seems to be looking for this feature.

Reviewed By: arthaud

Differential Revision: D26192496

fbshipit-source-id: 72cdf8b7b64ef590577f24496cc17de749b1ac27