make stubs integration test report errors
Summary:
We used to report model verification errors on the stderr, now we do so on stdout due to changes we've made to make the errors prettier and consistent with type checking.

This meant that the stubs integration test started not reporting them to the user, so let's bring that back.

Reviewed By: dkgi

Differential Revision: D26424445

fbshipit-source-id: 5baf5ee941db257680b8fb01b5f9a49c72d340bd