split leaf names from the simple feature set
Summary: We do not need to have an over-under set abstract domain for leaf names and can simple use a set. Also, we rarely need to update the leaf name set when propagating taint, hence it should be better to move it in its own set, in terms of performance. A downside is that we lose the `on_all_flows` annotation, but this is not used by SAPP anyway.

Reviewed By: fahndrich

Differential Revision: D26446721

fbshipit-source-id: f31dcfaef6b06dc86aba50abd340a44e2375479c