implement tito positions with a topped set abstract domain
Summary: This patch moves the taint-in-taint-out positions from the `Features.SimpleSet` into its own `Features.TitoPositionSet`, implemented with a topped set abstract domain. This should bring performance benefits due to keeping a limited amount of positions (here, 30) instead of blowing up. Another downside of the previous approach was that `SimpleSet` is an over and under approximation, but we never need the under approximation for tito positions. Finally, this should make the stripping of tito positions at call site faster since we don't need to iterate over all features.

Reviewed By: fahndrich

Differential Revision: D26386088

fbshipit-source-id: 6df7972574069cb193de0094f4063a06b2d3cbf2