Add code to inline decorator with one wrapper function.
Summary:
We have to populate captures and nesting_defines so that typechecking proceeds correctly. Otherwise, the call graph will end up wrong and we won't get the benefit of inlining.

Design choice: I'm inlining both the original function and the wrapper function, and wrapping them in a function that has the signature of the wrapper function.

```
with_logging
def foo(x: int) -> None:
  print(x)

# becomes (A)

    def foo(y: str) -> None:
      def original_function(z: str) -> None:
        print(z)

      def wrapper(y: str) -> None:
        __test_sink(y)
        original_function(y)

      return wrapper(y)

# not (B)

    def foo(y: str) -> None:
      def original_function(z: str) -> None:
        print(z)

      __test_sink(y)
      return original_function(y)
```

This has several benefits over inlining the original function into the wrapper function (B):
+ No scope conflicts in case the wrapper function has a nested function with the same name as the original function.
+ Flexibility when tackling decorators with multiple wrapper functions, like when we choose between an async and a non-async version based on a predicate. We wouldn't be able to handle that with (B)
+ Flexibility when adding code that is in the decorator outside the wrapper function. (Not yet handled.)

For the same reason, we don't inline the original function entirely and remove any nested functions. That would lead to scope conflicts, etc.

Reviewed By: arthaud

Differential Revision: D26402132

fbshipit-source-id: 4b921dddbb9ac1caf8f567a2e2c39c0f918cda8b