Add constraint-solver for variadic tuples.
Summary:
The solver stores intervals for variadic tuples. We go for a lattice that just has `Top`, `Bottom`, and tuples. The interval is just a single bound. If we get unequal bounds, we don't try to compute the `join` or `meet`.

(See code docstring for more details.)

Yet to hook this into `solve`.

Remaining work:
Yet to handle bounds that contain other variadics or even unary type variables.

Reviewed By: grievejia

Differential Revision: D26503728

fbshipit-source-id: 62a4035494ccdab9d6430d525ea7cae7062e04c5