Add variant for `Concatenation` in `OrderedTypes`
Summary:
This represents a concatenation of  a prefix of non-variadics, an unpacked variadic, and a suffix of non-variadics.

No functionality yet. This just fixes the type errors that pop up.

Note that we don't yet represent concatenation of multiple variadics or of unbounded tuples. In the future, this will be a prefix of non-variadics, an unpacked item, possible other items till we reach the last unpacked item, followed by a suffix of non-variadics.

The idea is that the non-variadic prefix and suffix can be matched and solved for directly. It's the middle part that requires special attention.

Reviewed By: grievejia

Differential Revision: D26503732

fbshipit-source-id: 9b5e340aaa72663c2c71d17a0b123ee90773c1cb