Add tests for `extract_partial_solution` for variadic tuples.
Summary:
This follows automatically from the implementation.

It checks what happens when we solve for a few out of the variables we have. For example, say we have intervals on two variadics: Ts has the upper bound Ts2 and Ts2 has the lower bound Ts.

`extract_partial_solution [Ts]` will return the solution that `Ts = Ts2`. It will also instantiate this partial solution in the remaining intervals: Ts2 has the lower bound Ts = Ts2.

Reviewed By: grievejia

Differential Revision: D26520889

fbshipit-source-id: ba7c30fdd7df5a5abdb740db25104bd7861297ce