Resolve `*args: *Ts` when typechecking a signature.
Summary:
The callable returned by `AnnotatedCallable.create_overload_without_applying_decorators` is used to check for existing type variables. If we don't implement this here, then it complains that the `Ts` in the return type was not found in the parameter types.

`TypeCheck.check_parameter_annotations` stores the annotation for each parameter of a define.

Reviewed By: grievejia

Differential Revision: D26653216

fbshipit-source-id: 402cd1515efa45f90dc9ca18917b9d2e063b79ef