Support fixme threshold in fixme-single upgrade command
Summary:
Really wanted this functionality when trying to do a custom upgrade for f3/share to resolve their timeouts.

They had a few files with a lot of type errors and I preferred to add a file-level suppression above a threshold to avoid making the code unreadable with comments.

We already support this flag in many of our codemods, adding it here too.

Reviewed By: pradeep90

Differential Revision: D26755751

fbshipit-source-id: 67f72e947626e11c5571b5e74d57ac327d70640d