Use resolved callee base as arguments of inverse operator
Summary:
Remove exponential time handling of inverse operators in type check.

We were re-resolving the base of calls in signature select, in order to get the type of that base when it's used as an argument.

This resulted in long chains of operators that get reversed to grow exponentially in runtime.

For example, in order to resolve `a >> b`, we first resolve the base `a` and then resolve the call `a.__rightshift__(b)`.

If we end up inverting the rightshift operator, we try resolving `b.__rrightshift__(a)`, but re-resolve `a` as an argument.

This means a chain of `a >> b >> c >> d` will resolve `d` once, `c` twice, `b` four times, `a` eight times, ... :'(

This diff saves the resolved base of the callee - which we have access to in forward expression logic already. For callee `a.__rightshift__`, that means we're saving the resolved type of `a` and passing this into `forward_callable`, so we can use the resolved type of `a` when using it as an argument.

----

f3/share in particular was hanging because it contains a lot of files with heavily chained rshift operators, and each element itself was often a very nested expression that is not cheap to resolve.
ie.
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/f3/merger/tests/test_merging.py?commit=ccc0bd51fe0d&lines=1263
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/f3/share/features/ads/gff_batch/GFF_C2607_ADS_ENTITY_EQUIVALENCE_KEY_SUM_7_USER_AGE_AD_CREATIVE_OPTIMIZATION_VIDEO_SUBTITLE_OPTVAL_350002_COUNTER_IMPRESSION_EXPLODED.py?commit=ccc0bd51fe0d&lines=34
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/f3/share/features/ads/gsf_tmp/gsf_float_user_non_dpa_scrolling_rates_page_type_19_agg_range_7.py?commit=ccc0bd51fe0d&lines=904
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/f3/share/features/ads/events/event_rtu_ads_conversion.py?commit=ccc0bd51fe0d&lines=22

 ---

Related diffs for context:
Adding pre-resolved args: D21025126 (https://github.com/facebook/pyre-check/commit/8c18c8d2a7376c34179bad74e0fa96573b755ecc), D21026994 (https://github.com/facebook/pyre-check/commit/91fb706ba0bc229179c30341c12db3d68110c401)
Enabling inverse operators: D18828820 (https://github.com/facebook/pyre-check/commit/7f16261122c13a0b34d864f4fb004970c5abcb93)
 ---

Follow-ups:
- will simplify signature select to expect resolved arguments in all cases
- add back f3 deleted targets
- early-bump f3/share version to stop timeouts on master & fix decorated-related type errors

Reviewed By: grievejia

Differential Revision: D26737597

fbshipit-source-id: dd06d1fe61e15cd6857ee61e52c27582beabfe2b