Use attribute and non attribute variant for forward_callable
Summary:
Feedback from D26737597 (https://github.com/facebook/pyre-check/commit/33993023b04edab0930efcdbc06ec11186245b8e) -- using a variant as input into forward_callable rather than a series of callable-related inputs.

This way, we ensure that resolved_callable_base is only passed in when the callee is an attribute, and hopefully make this function interface more readable with fewer re-matches in the body to determine if we need to try a reverse operator, etc.

Reviewed By: grievejia

Differential Revision: D26802703

fbshipit-source-id: 52dac76aa5b634679707c9bd8fb83556fdc4e8ec