Resolve concatenation of tuples via unpacking.
Summary:
Problem: We used to resolve `(1, *(2, 3))` as `Tuple[int, unknown]`. This was because we correctly resolved `(2, 3)` as `Tuple[int, int]` and then dropped the ball on the `Starred` expression because there is no type to express a naked unpacked tuple.

With variadic tuples, we can handle different kinds of unpacked tuples.

Reviewed By: dkgi

Differential Revision: D26821042

fbshipit-source-id: 7aad420ca4b2775e32617077a4f1d0e965efff30