Check `is_meta` before extracting element.
Summary:
Problem: When given multiple types for `isinstance`, we check against each of them. In particular, for `is_instance(x, tuple_of_types)`, if the types are `Tuple[Type[int], ...]`, we extract the inner `int` and compare `x` against that.

If the types are `Tuple[Any, ...]`, we should just give up on the check and raise an error. However, I extracted the inner parameter and then did `Option.some_if`.

Reviewed By: grievejia

Differential Revision: D26839732

fbshipit-source-id: c408cdc52013b286055b777e2865022abd80a271