Remove uses of `map2_exn`.
Summary: Problem: Now that we allow `(1, *(2, 3)`, that means the number of elements in the tuple expression may not match the number of elements in the tuple type `Tuple[int, int, int]`. Weaken literals was assuming that the number of elements in the expression, actual type, and expected type would match.

Reviewed By: grievejia

Differential Revision: D26840404

fbshipit-source-id: c9e0708f7c520297e068798073a90cc2f017a802