Updated guidelines for building Pyre from source (#387)
Summary:
There are some missing steps in the docs for building and testing pyre from source. As per my understanding, the internal build and testing system for Pyre is different from the one in open source.

The changes, as discussed with gbleaney, aim to simplify contributing to Pyre by documenting the extra steps/ possible issues contributors might face while setting up the development environment from source.

The reason behind each of the changes are as follows:

1. As I understand, `run-python-tests.sh` is a replacement for the older `make python_tests` command. I have noticed that running the script directly gives some "missing modules" errors, which are resolved by  installing the dependencies via `requirements.txt`. Hence, documenting this step might be a good idea.
2. After completing all the installation steps, the only way to run pyre is via `client/pyre.py` wrapper from the root directory. However, trying to run it directly gives `relative import` errors, which are resolved only by invoking it as a module. This makes it a frustrating task to call Pyre from anywhere other than the root directory. Adding the root to `PYTHONPATH` and making an alias is a workaround to that.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/387

Reviewed By: grievejia

Differential Revision: D26781005

Pulled By: gbleaney

fbshipit-source-id: 4c47a9bf3c89e28196130e2823e94661cba17837