generate a model verification error for models on inherited properties
Summary:
We do not materialize methods in child classes if they are only defined in the parent class. We currently generate a model verification error if a user tries to model the method on the child class in that situation.

We have the same problem for properties, but we currently do not error out. This patch adds a model verification error when modeling an inherited property that is only defined in the parent class.

Reviewed By: dkgi

Differential Revision: D26897124

fbshipit-source-id: a0f55b15fdbe76f198a1a037595d4e5cf0bd9cc8