Add synthetic local variables for `args` and `kwargs` used within the decorator.
Summary:
Problem: We tighten the signature of the wrapper function if it always passes on `*args, **kwargs` to the original function. However, we need to catch any flows to sinks within the wrapper (such as `log(args)`). But we no longer have the full `args` or `kwargs` available after making the signature match the signature of the original function.

So, I'm creating two synthetic local variables `__args` and `__kwargs` that contain all the arguments. This way, we will continue to catch flows to sinks within the wrapper.

Design choice:

I'm storing all arguments in both `__args` and `__kwargs`. This may lead to false positives since a named parameter would end up in exactly one or the other based on the call. However, we can't tell that at static analysis time, so this is a compromise.

Not trying to be too precise about the keys for the `__kwargs` dictionary. Just using `Expression.show parameter` for now. I think the most important thing is to ensure that there is a flow from each parameter into the `__kwargs` dictionary. We can revisit this later if we need to be more precise.

Reviewed By: arthaud

Differential Revision: D26982076

fbshipit-source-id: b8f919849c11fccb01ce3aa8b908360e6a056d82