Add helper function for tightening the signature of a broad decorator wrapper.
Summary:
Problem: If a decorator has type `def wrapper(*args, **kwargs)` and we simply use that as the signature of the decorated function, then we will lose precise taint information about each parameter.

As a heuristic, if the wrapper always calls the original function as `original(*args, **kwargs)`, then make it have the same parameters as the original function.

For now, I'm adding a synthetic local variable to hold the full set of arguments, since that would need to be used within the decorator itself. (Will handle edge cases like `kwargs`, etc. in upcoming diffs.)

Reviewed By: arthaud

Differential Revision: D26982074

fbshipit-source-id: 42e577097139c4d545e826b5e6f99a9d746598de