log slow methods every time we reach a threshold
Summary:
The current implementation logs methods that take more than 60 seconds to analyze.
The idea of this patch is to log again at 120s, 240s, 580s etc.
This will allow us to catch functions that make the analysis timeout.
Note that we already log the slowest methods at the end of the analysis, but this is not useful when the analysis timeouts.

Reviewed By: dkgi

Differential Revision: D27031888

fbshipit-source-id: e038c8322f8e8314ff95ac67682e5d3cb7afc1af