Updated installation docs
Added an extra instruction for pointing Pyre to the location of the compiled binary. After a fresh installation from source, after following the previously written steps, it always seems to want to know the location of `pyre.bin` before running. This change takes care of that and ensures the user won't have to figure this out on their own.