Updated installation docs (#394)
Summary:
Added an extra instruction for pointing Pyre to the location of the compiled binary. After a fresh installation from source and following the previously written steps, it always seems to want to know the location of `pyre.bin` before running. This change takes care of that and ensures the user doesn't have to figure this out on their own.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/394

Reviewed By: dkgi

Differential Revision: D27268601

Pulled By: gbleaney

fbshipit-source-id: b393362c0abfe66d3b45f4162dbfebf393b1ee8e