Pass well-typed config to analysis initialize
Summary:
While working on hooking up a proper initialization step up
to typeInferenceAnalysis, I realized that the top-level logic in
staticAnalsyis.ml was doing something bad: it was converting the
raw config Configuration.StaticAnalysis.t into Json before passing it
to the ANALYZER.init function of the Analysis modules.

This is bad for two reasons:
- firstly, it was dropping type safety at a very high layer of the code
  and using json, which basically amounts to passing dynamically-typed
  data around. There are reasons we might want to do this, but none of them
  seem to apply here
- secondly, it was coupling to the taint case by passing exactly the
  data taint needs from a top-level function in what is ostensibly
  analysis-implementation-agnostic code

This commit fixes the main problem without rewiring too much actual code by
just pushing the json cast into a local function of taintAnalysis.ml. It would
be better to eventually rewire all of the taint code to use a well-typed
config object of `Configuration.StaticAnalysis.t` type, but that might take a
while and for now my concern is just to ensure that the top-level generic
code is actually generic.

Reviewed By: dkgi

Differential Revision: D27345792

fbshipit-source-id: 0bf2f6ca0faac09de8d253ee870fa4411a994e44