Hook inference logic up to interprocedural analysis
Summary:
This commit hooks the local type inference logic from `Analysis.Inference.run`
to the `analyze` function used to drive interprocedural analysis in
`TypeInferenceaAnalysis`.

We modify the `result` type to be a list of `AnalysisError.Instantiated.t`,
which matches what we use to print json in the existing `pyre infer` command,
and update the `externalize` function to use the same format.

This should mean that we can update the client to use `analyzeCommand` rather
than `inferCommand` fairly easily; we'll still need to change from consuming
stdout to consuming a directory of json data as `pyre analyze` does but the
low-level data will be the same at first; we can move away from using
`AnalysisError` later.

Reviewed By: dkgi

Differential Revision: D27419358

fbshipit-source-id: 09344a2edcccdb008df73680b69ea08e2f92a7e4