Refactor one-define inference into `run_local`
Summary:
As we migrate type inference into the interprocedural static analysis
framework, one of the changes is that instead of scheduling type inference
on a module at a time the static analysis scheduler will infer one callable
at a time.

As a result, we want as much of the type inference logic as possible to
operate on a single `Define.t Node.t` at a time. That logic was buried in
the `check` inline function in master, in this diff we extract as much
as possible into `run_local`.

There's a little extra work introduced here: because I don't
want to change the output of the existing `pyre infer` command but I
also don't want to lose any of the error processing logic, I now
run the `join_at_source` and `sort` logic twice - once inside the
`run_local` and again in the old `run` command, whose output should
be unchanged.

Reviewed By: dkgi

Differential Revision: D27449232

fbshipit-source-id: 914a373f362d7a30f67c48a1921a6eb62f225100