Make pyre_dump_cfg logic match documentation
Summary:
I realized when trying out all the debug tools that we
had actually removed the pyre_dump_cfg debug logic. This diff
adds it back - if we come across a `pyre_dump_cfg()` call in the
body of a function, then we'll pretty-print the control flow graph
for that function body.

At a suggestion in diff review, we dump the control flow graph in dot format (as a single line) so that it's easy to paste into a graphviz tool

I also added checks of pyre_dump_cfg and pyre_dump_locations
to the unit tests where pyre_dump was tested. The relevant logic
didn't actually change, but I came across it when checking whether
there were unit tests to update and it seems like we might as well
check for all three.

Reviewed By: dkgi

Differential Revision: D27403437

fbshipit-source-id: 1f2453da22d728ee645c80f7ed182cf4ffea545b