Ignore decorators that can't be inlined.
Summary:
Problem: We skip inlining if decorators have no wrapper function or are just generators. However, if there are multiple decorators and one of them can't be inlined, we give up on the entire function.

Since there are quite a few irrelevant decorators that can't be inlined, we end up missing out on quite a few callsites.

Instead of just defaulting to the function without any decorators, i.e., ignoring all decorators, we can just ignore those decorators that can't be inlined.

Reviewed By: arthaud

Differential Revision: D27490720

fbshipit-source-id: 431816486c8525b55b295bd7478345c26270a220