Implement TypeAlias from typing_extensions instead of typing
Summary:
TypeAlias was moved from typing to typing_extensions.

This breaks our support of TypeAlias. If you attempt to annotate a type alias as such:

```
X: TypeAlias = int
```

You'll get a type error:
Incompatible variable type [9]: X is declared to have type TypeAlias but is used as type ....

Fix this so we can still declare aliases with `TypeAlias`, esp. useful when Pyre is trying to interpret the RHS as a global that needs a type annotation.

Reviewed By: dkgi

Differential Revision: D27527037

fbshipit-source-id: dbc7b99785fb6bdb1501b153420b85645aada2c6