upgrade pysa module to use graphql3
Summary:
Since the pyse module is used in distillery and distillery is already on graphql3, we have to upgrade the fbcode code to graphql3 and then we can bring this module back in distillery.

This diff upgrades to graphql3.

Since this is an open source thing, do I need to do anything extra? Please let me know. Thanks!

Reviewed By: gbleaney

Differential Revision: D27481062

fbshipit-source-id: d15ec38be57ab3669aa06f9a81e425a8f33f72bc