fix missing postconditions due to collapsing [1/n]
Summary:
We can emit invalid traces when the taint is collapsed between 2 iterations.

# Example
For instance, imagine that `foo` calls `bar`.
At iteration 1, `bar` has a model with a source `LocalResult[a] -> Origin(...) -> UserControlled -> ...`.
At iteration 2, `foo` propagates the model for `bar` into `CallSite(callees=[bar], port=result[a]) -> UserControlled -> ...`.
Now at iteration 3, `bar` propagates more sources, and reaches the max width, hence its taint is `LocalResult -> Origin(..) -> UserControlled -> ...` (the path is empty).
The problem is that `foo` still refers to `result[a]` from iteration 1. This will be kept until we reach the fixpoint by the property of the fixpoint (we always widen the old model and the new model between iterations).

# Current solution
To counter this problem, we do have a trick when we emit our traces: we expand override targets, and filter callables that do not have a taint for a specific path. The problem is that this is not done correctly: it checks if there is any child for that path (i.e, there could be `LocalResult[a][b]`) before filtering out. In this patch, we fix this part by only checking for `LocalResult[a]`

# Path forward
There are still other problems not addressed here: we don't check if the taint found in the callee correspond to the same tain kind (i.e, `Sources.t` or `Sinks.t`).
I also think that this pruning needs to be propagated: In the previous example, the taint in `foo` might be propagated to other functions, so we would also need to prune invalid ports in those. I think we basically need a decreasing fixpoint to remove all invalid taint. I don't think this work is high priority/impact unfortunately. If I understand correctly, those invalid flows just generate warnings in the zoncolan UI.

Reviewed By: fahndrich

Differential Revision: D27529136

fbshipit-source-id: e998ced3c2fe6e74c78f1c22467209d6c9e43f3e