Fix pyre statistics nested scope tracking
Summary:
Pyre statistics was losing annotated parameter information for functions following a nested class definition.

Initially thought we were failing to visit functions following the nested class definition but it turns out we are manually tracking whether or not we are inside a class definition and using this to count `self` params as annotated. We need to use a counter here rather than a boolean in order to handle going in and out of nested classes.

Note: will need to land, bump client, and then run backfills.

Reviewed By: dkgi

Differential Revision: D27579570

fbshipit-source-id: 8ddee965f67f140037822c734149d3b3cee3031b