Added a new command for starting a pysa language server (#405)
Summary:
This change would create a `pysa-language-server` to connect to Pysa's language server. How should we handle versioning of commands here? Pysa server was introduced in v2 folder, so should we assume that the command works in v2?

The `stop` command seems to be generic for all running servers (it iterates through socket files) -- should there be additional command for stopping Pysa's server?

Pull Request resolved: https://github.com/facebook/pyre-check/pull/405

Reviewed By: grievejia

Differential Revision: D27567709

Pulled By: gbleaney

fbshipit-source-id: c54aa196c7daaf85e9ea385d103fd87154a45a70