Add type to represent arbitrary literal string.
Summary:
Problem: We want to specify that a given variable accepts only literal strings. We can't do that with `Literal["foo"]` since that requires a specific string.

So, I'm using `Literal[str]` to represent literal strings. The change turned out to be relatively painless because it's encapsulated within `Literal (String _)`. We didn't have to change the code for weakening or joining string literals.

Can add this for integer literals, etc. in the future if there is demand.

Design choice:
+ Not adding new variants `type literal = | String of string | AnyStringLiteral | ...` because that would duplicate each existing variant.

Alternatives:
+ Considered making this work: `Literal[T], where T <: str`. However, that's not quite right. `Literal[]` takes a value, not a type. `T <: str` means that `T = Literal["foo"]`, etc., which would give us `Literal[Literal["foo"]]`. We could flatten this and treat it as `Literal["foo"]` but that might confuse users reading error messages.
+ Ideally, we would express this as `Literal[Any] & str` after implementing `Literal[Any]` and intersection types. That would also extend to `Literal[Any] & int`, etc. I tried representing that as a Protocol with two base classes, `Literal` and `str`, but `Literal` is not a valid base class at runtime.
+ Alternative for the future: Allow `Literal[<type>]`. This would let us write `d: Literal[Dict[str, str]] = {"foo": "bar"}`. Not sure how useful this would be. Leaving this aside for now.

Going with `Literal[str]` instead of `pyre_extensions.LiteralString` to reuse the familiar type and avoid unnecessary new dependencies.

Reviewed By: dkgi

Differential Revision: D27615818

fbshipit-source-id: 2069b4e3aeaa5926e0d18a2389bbb5fc8c423263